"Magma" fixes (#57)
* DFCIR, DFCxx: added bit extraction and concatenation operations; DFCxx: ref-based API replaced with ptr-based API.
* DFCIR: added dfcir.bits and dfcir.cat to graph construction; graph construction API was simplified.

* DFCIRToFIRRTLPass: added BitsOp and CatOp conversion procedures.
* DFCIRToFIRRTLPass: DFCIR-FIRRTL type conversion simplified & removed copy-paste.
* DFCIRToFIRRTLPass: removed code which is supposed to appear in a separate MR.

* DFCxx: bitrange boundaries have uint16_t types now.
* DFCxx: changed FixedType and FloatType to use uint16_t instead of uint8_t.
* DFCxx: KernStorage::addVariable API changed to allow custom comparators, DF-type comparison fixed.
* DFCxx: Graph::addNode now skips adding the same variable twice.
* DFCxx: multiplexers now unconditionally produce streams instead of copying the driver of the first input.

* DFCIRASAPSchedulerPass: inefficient channel visiting was fixed.

* examples: Magma-example introduced (magma_encoder for now).
* examples/magma: added XOR_INT config value; replaced the wrongly placed addition with XOR.

* FIRRTLStubsGeneratorPass: added some of the bitwise operators.

* DFCIR, Scheduling: changed graph channels API to use vectors instead of unordered sets; fixed overall module delay calculation.

* MagmaEncoder: implementation was fixed and a GOST testbench was added.
* MagmaDecoder: implementation was added and a GOST testbench was added.

---------

+++ src/model/dfcxx/include/dfcxx/kernel.h
@@ -51,11 +51,11 @@ protected:
   Constant constant;
   Control control;
 
-  DFType dfUInt(uint8_t bits);
+  DFType dfUInt(uint16_t bits);
 
-  DFType dfInt(uint8_t bits);
+  DFType dfInt(uint16_t bits);
 
-  DFType dfFloat(uint8_t expBits, uint8_t fracBits);
+  DFType dfFloat(uint16_t expBits, uint16_t fracBits);
 
   DFType dfBool();
